Make EO tokens be consistently followed by newlines
In this commit, the completion functions now all return an expression
either of the form "EO\n<completions>" or
"<options>\nEO\n<completions>", making the EO token always be followed
by a newline, and be preceded by a newline if necessary; previously some
functions omitted one or both newlines.  The behavior of these functions
should now be consistent.