Allow regex metacharacters in user input
Previously, several of the bash completion functions used grep and sed
commands which included user input in the matching expression.
Unfortunately, if the user input contained regex metacharacters or other
characters which would be recognized by the command being used as having
special meaning, the command would fail to have its intended effect.
This commit attempts to fix that by replacing those commands with a
utility function which uses awk in such a way that no special characters
will be interpreted from the user-input string.